Move debounced to functional utils.
